Move LogsComponent into new LogsPage directory
