+++ src/app/home/pages/LogsPage/components/LogBody.tsx
@@ -12,8 +12,8 @@ import LogItem from './LogItem';
 import { connect } from 'react-redux';
 import { Spinner } from '@patternfly/react-core';
 import { FunctionComponent } from 'react';
-import { LogActions } from '../duck';
-import { IPlanLogSources } from '../../../client/resources/discovery';
+import { LogActions } from '../../../../logs/duck';
+import { IPlanLogSources } from '../../../../../client/resources/discovery';
 
 interface IProps {
   isFetchingLogs: boolean;
