+++ src/app/home/pages/LogsPage/components/LogItem.module.scss
