+++ src/app/home/pages/LogsPage/components/LogItem.tsx
