+++ src/app/home/pages/LogsPage/index.ts
@@ -1,0 +1,1 @@
+export * from './LogsPage';
