+++ src/app/home/pages/index.ts
@@ -1,3 +1,4 @@
 export * from './ClustersPage';
 export * from './PlansPage/PlansPage';
 export * from './StoragesPage';
+export * from './LogsPage';
